SpyderPdb: initial organization for SpyderPdb.
